put handlers into cmswww
