+++ politeiawww/cmswww.go
@@ -7,6 +7,7 @@ import (
 	cms "github.com/decred/politeia/politeiawww/api/cms/v1"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/util"
+	"github.com/gorilla/mux"
 )
 
 // handleInviteNewUser handles the invitation of a new contractor by an
@@ -56,6 +57,77 @@ func (p *politeiawww) handleRegisterUser(w http.ResponseWriter, r *http.Request)
 	}
 
 	// Reply with the verification token.
+	util.RespondWithJSON(w, http.StatusOK, reply)
+}
+
+// handleNewInvoice handles the incoming new invoice command.
+func (p *politeiawww) handleNewInvoice(w http.ResponseWriter, r *http.Request) {
+	// Get the new proposal command.
+	log.Tracef("handleNewInvoice")
+	var ni cms.NewInvoice
+	decoder := json.NewDecoder(r.Body)
+	if err := decoder.Decode(&ni); err != nil {
+		RespondWithError(w, r, 0, "handleNewInvoice: unmarshal", www.UserError{
+			ErrorCode: www.ErrorStatusInvalidInput,
+		})
+		return
+	}
+
+	user, err := p.getSessionUser(w, r)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleNewInvoice: getSessionUser %v", err)
+		return
+	}
+
+	reply, err := p.processNewInvoice(ni, user)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleNewInvoice: processNewInvoice %v", err)
+		return
+	}
+
+	// Reply with the challenge response and censorship token.
+	util.RespondWithJSON(w, http.StatusOK, reply)
+}
+
+// handleInvoiceDetails handles the incoming invoice details command. It fetches
+// the complete details for an existing invoice.
+func (p *politeiawww) handleInvoiceDetails(w http.ResponseWriter, r *http.Request) {
+	// Add the path param to the struct.
+	log.Tracef("handleInvoiceDetails")
+	var pd cms.InvoiceDetails
+
+	// get version from query string parameters
+	err := util.ParseGetParams(r, &pd)
+	if err != nil {
+		RespondWithError(w, r, 0, "handleInvoiceDetails: ParseGetParams",
+			www.UserError{
+				ErrorCode: www.ErrorStatusInvalidInput,
+			})
+		return
+	}
+
+	// Get proposal token from path parameters
+	pathParams := mux.Vars(r)
+	pd.Token = pathParams["token"]
+
+	user, err := p.getSessionUser(w, r)
+	if err != nil {
+		if err != ErrSessionUUIDNotFound {
+			RespondWithError(w, r, 0,
+				"handleInvoiceDetails: getSessionUser %v", err)
+			return
+		}
+	}
+	reply, err := p.processInvoiceDetails(pd, user)
+	if err != nil {
+		RespondWithError(w, r, 0,
+			"handleInvoiceDetails: processInvoiceDetails %v", err)
+		return
+	}
+
+	// Reply with the proposal details.
 	util.RespondWithJSON(w, http.StatusOK, reply)
 }
